prototype layout
