+++ Prototype/layout.txt
@@ -1,0 +1,61 @@
+<Header>
+  <Nav>
+    <Row>
+      <LogoIcon />
+      <Searchbar /> fetch of type post
+      <LoginIcon /> ---> <Login /> fetch user from database
+      <CartIcon /> ---> <Cart />
+    </Row>
+    <Row>
+      <ShopCategories />
+    </Row>
+  </Nav>
+</Header>
+
+<Homepage>
+  <Section>
+    <Cover>
+      <Hero>
+        <Title />
+        <Description />
+        <CallToActionBtn /> ---> <Shop />
+      </Hero>
+    </Cover>
+  </Section>
+
+  <Section>
+    <GridShopCategories /> <Category />
+  </Section>
+
+  <Section>
+    <Footer />
+  </Section>
+</Homepage>
+
+
+<Shop>
+  1. with no category
+  2. with a category
+
+  const get_url = category ? `/api/${category}` : `/api/allProducts`
+
+  fetch(get_url)
+
+  <Filter />
+    1. num of items
+    2. category main
+    3. body location sub
+    4. company
+    5. price
+
+  fetch all products by category
+  <ProductList numOfItems={10}> items.map(return <Product />)
+    <Product />
+    <Pagination />
+  </ProductList>
+</Shop>
+
+fetch post submit
+<SideCart />
+<FullCart />
+
